Refactor coding agent workflow and namespaces
This commit introduces a new namespace `MCP.Host.Agents.CodingAgent` to encapsulate the coding agent workflow and related classes, improving organization and clarity. The `RunAsync` method in `CodingAgentWorkflow` now utilizes a new `WorkflowParameter` for better parameter handling.

Additionally, the `CodingAgentBackgroundService`, `ICodingAgentWorkflowStore`, and `CodingAgentWorkflowStore` classes have been updated to reflect the new namespace structure. Event-related classes `AgentOrchestrationEvents` and `CommonEvents` have also been moved to the new namespace.

These changes enhance modularity, readability, and maintainability by clearly defining the responsibilities and interactions of the coding agent components.
