+++ src/MCP.Host/Agents/CodingAgent/Events/CommonEvents.cs
@@ -1,0 +1,9 @@
+﻿namespace MCP.Host.Agents.CodingAgent.Events;
+
+public static class CommonEvents
+{
+    public static readonly string UserInputReceived = nameof(UserInputReceived);
+    public static readonly string UserInputComplete = nameof(UserInputComplete);
+    public static readonly string AssistantResponseGenerated = nameof(AssistantResponseGenerated);
+    public static readonly string Exit = nameof(Exit);
+}
