+++ src/MCP.Host/Agents/CodingAgent/WorkflowParameter.cs
@@ -1,0 +1,3 @@
+﻿namespace MCP.Host.Agents.CodingAgent;
+
+public record WorkflowParameter(Guid ChatId, string ConnectionId, string Requirement);
