+++ src/MCP.Host/Services/KernelFactory.cs
@@ -1,4 +1,4 @@
-using MCP.Host.Agents;
+using MCP.Host.Agents.CodingAgent.Steps;
 using MCP.Host.Plugins;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Agents;
